+++ src/ops/op.c
@@ -85,7 +85,7 @@ const tele_op_t *tele_ops[E_OP__LENGTH] = {
     &op_TR_POL, &op_TR_TIME, &op_TR_TOG, &op_TR_PULSE, &op_TR_P, &op_CV_SET,
     &op_MUTE, &op_STATE, &op_DEVICE_FLIP, &op_LIVE_OFF, &op_LIVE_O,
     &op_LIVE_DASH, &op_LIVE_D, &op_LIVE_GRID, &op_LIVE_G, &op_LIVE_VARS,
-    &op_LIVE_V, &op_PRINT, &op_PRT, &op_CV_GET,
+    &op_LIVE_V, &op_PRINT, &op_PRT, &op_CV_GET, &op_CV_CAL, &op_CV_CAL_RESET,
 
     // maths
     &op_ADD, &op_SUB, &op_MUL, &op_DIV, &op_MOD, &op_RAND, &op_RND, &op_RRAND,