+++ src/ops/op_enum.h
@@ -165,6 +165,8 @@ typedef enum {
     E_OP_PRINT,
     E_OP_PRT,
     E_OP_CV_GET,
+    E_OP_CV_CAL,
+    E_OP_CV_CAL_RESET,
     E_OP_ADD,
     E_OP_SUB,
     E_OP_MUL,